--- backend/core/tests/__init__.py
@@ -1,1 +1,0 @@
-# Core application tests
