--- backend/core/tests/test_security.py
@@ -1,481 +1,0 @@
-# backend/core/tests/test_security.py
-"""
-Comprehensive security test suite for LifePlace platform
-
-Run with: python manage.py test core.tests.test_security
-"""
-
-import json
-import hashlib
-import hmac
-from unittest.mock import patch, MagicMock
-
-from django.test import TestCase, RequestFactory, override_settings
-from django.contrib.auth import get_user_model
-from django.core.files.uploadedfile import SimpleUploadedFile
-from rest_framework.test import APIClient, APITestCase
-from rest_framework import status
-
-from core.utils.security import (
-    validate_email_format,
-    validate_password_strength, 
-    validate_file_upload,
-    sanitize_input,
-    LoginRateThrottle,
-    RegistrationRateThrottle
-)
-from core.utils.encryption import encrypt_data, decrypt_data, EncryptionService
-from core.utils.security_logging import (
-    SecurityLogger,
-    SecurityEvent,
-    SecurityEventType,
-    SecuritySeverity,
-    security_logger
-)
-from core.domains.payments.models import PaymentGateway
-from core.domains.users.services import UserService
-
-User = get_user_model()
-
-
-class SecurityUtilsTestCase(TestCase):
-    """Test security utility functions"""
-    
-    def test_email_validation(self):
-        """Test email format validation"""
-        # Valid emails
-        valid_emails = [
-            "test@example.com",
-            "user.name@domain.co.uk", 
-            "test+tag@example.org",
-            "123@example.com"
-        ]
-        
-        for email in valid_emails:
-            self.assertTrue(validate_email_format(email), f"Should accept valid email: {email}")
-        
-        # Invalid emails
-        invalid_emails = [
-            "invalid-email",
-            "@example.com",
-            "test@",
-            "test.example.com",
-            "",
-            "test@.com",
-            "test@example.",
-            "test@@example.com"
-        ]
-        
-        for email in invalid_emails:
-            self.assertFalse(validate_email_format(email), f"Should reject invalid email: {email}")
-    
-    def test_password_strength_validation(self):
-        """Test password strength validation"""
-        # Strong passwords
-        strong_passwords = [
-            "StrongP@ssw0rd123",
-            "MySecur3P@ssword!",
-            "C0mpl3x-P@ssw0rd"
-        ]
-        
-        for password in strong_passwords:
-            result = validate_password_strength(password)
-            self.assertTrue(result['is_valid'], f"Should accept strong password: {password}")
-            self.assertGreaterEqual(result['score'], 4)
-        
-        # Weak passwords
-        weak_passwords = [
-            "123",
-            "password",
-            "abc123",
-            "short",
-            "password123",
-            "qwerty"
-        ]
-        
-        for password in weak_passwords:
-            result = validate_password_strength(password)
-            self.assertFalse(result['is_valid'], f"Should reject weak password: {password}")
-    
-    def test_input_sanitization(self):
-        """Test input sanitization"""
-        test_cases = [
-            # (input, expected_contains, expected_not_contains)
-            ("<script>alert('xss')</script>Hello", "Hello", "<script>"),
-            ("<img src='x' onerror='alert(1)'>", "", "<img"),
-            ("Normal text", "Normal text", None),
-            ("Text with <b>bold</b> tags", "Text with bold tags", "<b>"),
-            ("", "", None)
-        ]
-        
-        for input_text, should_contain, should_not_contain in test_cases:
-            result = sanitize_input(input_text)
-            
-            if should_contain:
-                self.assertIn(should_contain, result)
-            if should_not_contain:
-                self.assertNotIn(should_not_contain, result)
-    
-    def test_file_upload_validation(self):
-        """Test file upload validation"""
-        # Create mock file objects
-        def create_mock_file(name, content_type, size):
-            mock_file = MagicMock()
-            mock_file.name = name
-            mock_file.content_type = content_type
-            mock_file.size = size
-            return mock_file
-        
-        # Test allowed file types
-        pdf_file = create_mock_file("document.pdf", "application/pdf", 1024*1024)
-        result = validate_file_upload(pdf_file, allowed_types=["application/pdf"], max_size_mb=10)
-        self.assertTrue(result['is_valid'])
-        
-        # Test disallowed file types
-        exe_file = create_mock_file("malware.exe", "application/x-executable", 1024*1024)
-        result = validate_file_upload(exe_file, allowed_types=["application/pdf"], max_size_mb=10)
-        self.assertFalse(result['is_valid'])
-        
-        # Test file size limits
-        large_file = create_mock_file("huge.pdf", "application/pdf", 100*1024*1024)
-        result = validate_file_upload(large_file, max_size_mb=10)
-        self.assertFalse(result['is_valid'])
-
-
-class EncryptionTestCase(TestCase):
-    """Test encryption functionality"""
-    
-    def setUp(self):
-        """Set up test environment"""
-        self.encryption_service = EncryptionService()
-    
-    def test_encryption_decryption(self):
-        """Test basic encryption and decryption"""
-        test_data = {
-            "stripe_secret": "sk_test_123456789",
-            "publishable_key": "pk_test_987654321"
-        }
-        
-        # Encrypt
-        encrypted = encrypt_data(test_data)
-        self.assertIsInstance(encrypted, str)
-        self.assertNotEqual(encrypted, str(test_data))
-        
-        # Decrypt
-        decrypted = decrypt_data(encrypted)
-        self.assertEqual(decrypted, test_data)
-    
-    def test_string_encryption(self):
-        """Test string encryption"""
-        test_string = "sensitive_api_key_12345"
-        
-        encrypted = encrypt_data(test_string)
-        decrypted = decrypt_data(encrypted, return_json=False)
-        
-        self.assertEqual(decrypted, test_string)
-    
-    def test_empty_data_handling(self):
-        """Test handling of empty/None data"""
-        self.assertEqual(encrypt_data(None), "")
-        self.assertEqual(decrypt_data(""), {})
-        self.assertEqual(decrypt_data("", return_json=False), "")
-    
-    def test_invalid_decryption(self):
-        """Test handling of invalid encrypted data"""
-        invalid_data = "invalid_encrypted_data"
-        result = decrypt_data(invalid_data)
-        self.assertEqual(result, {})  # Should return empty dict instead of failing
-
-
-@override_settings(CELERY_TASK_ALWAYS_EAGER=True)
-class SecurityLoggingTestCase(TestCase):
-    """Test security logging functionality"""
-    
-    def setUp(self):
-        """Set up test environment"""
-        self.factory = RequestFactory()
-        self.user = User.objects.create_user(
-            email="test@example.com",
-            password="TestPass123!",
-            role="CLIENT"
-        )
-        self.security_logger = SecurityLogger()
-    
-    def test_log_event_creation(self):
-        """Test that security events are logged to database"""
-        request = self.factory.post('/login/')
-        request.META['REMOTE_ADDR'] = '192.168.1.1'
-        request.META['HTTP_USER_AGENT'] = 'Test Agent'
-        
-        # Log an event
-        event = self.security_logger.log_event(
-            SecurityEventType.LOGIN_SUCCESS,
-            "Test login success",
-            request=request,
-            user=self.user,
-            severity=SecuritySeverity.LOW
-        )
-        
-        # Verify event was created
-        self.assertIsNotNone(event)
-        self.assertEqual(event.event_type, SecurityEventType.LOGIN_SUCCESS)
-        self.assertEqual(event.user, self.user)
-        self.assertEqual(event.ip_address, '192.168.1.1')
-        
-        # Verify it was saved to database
-        saved_event = SecurityEvent.objects.get(id=event.id)
-        self.assertEqual(saved_event.description, "Test login success")
-    
-    def test_login_success_logging(self):
-        """Test login success logging"""
-        request = self.factory.post('/login/')
-        request.META['REMOTE_ADDR'] = '192.168.1.1'
-        
-        self.security_logger.log_login_success(request, self.user)
-        
-        # Verify event was logged
-        events = SecurityEvent.objects.filter(event_type=SecurityEventType.LOGIN_SUCCESS)
-        self.assertEqual(events.count(), 1)
-        self.assertEqual(events.first().user, self.user)
-    
-    def test_login_failure_logging(self):
-        """Test login failure logging"""
-        request = self.factory.post('/login/')
-        request.META['REMOTE_ADDR'] = '192.168.1.1'
-        
-        self.security_logger.log_login_failure(request, "test@example.com", "Invalid password")
-        
-        # Verify event was logged
-        events = SecurityEvent.objects.filter(event_type=SecurityEventType.LOGIN_FAILURE)
-        self.assertEqual(events.count(), 1)
-        event = events.first()
-        self.assertEqual(event.username, "test@example.com")
-        self.assertIn("Invalid password", event.description)
-    
-    def test_risk_score_calculation(self):
-        """Test risk score calculation"""
-        request = self.factory.post('/login/')
-        
-        # Test high risk event
-        high_risk_event = self.security_logger.log_event(
-            SecurityEventType.SUSPICIOUS_ACTIVITY,
-            "Suspicious activity detected",
-            request=request,
-            severity=SecuritySeverity.HIGH
-        )
-        
-        self.assertGreaterEqual(high_risk_event.risk_score, 50)
-        
-        # Test low risk event
-        low_risk_event = self.security_logger.log_event(
-            SecurityEventType.LOGIN_SUCCESS,
-            "Normal login",
-            request=request,
-            user=self.user,
-            severity=SecuritySeverity.LOW
-        )
-        
-        self.assertLessEqual(low_risk_event.risk_score, 30)
-
-
-class PaymentGatewayEncryptionTestCase(TestCase):
-    """Test payment gateway configuration encryption"""
-    
-    def test_gateway_config_encryption(self):
-        """Test that payment gateway configs are encrypted"""
-        config_data = {
-            "secret_key": "sk_test_123456789",
-            "publishable_key": "pk_test_987654321"
-        }
-        
-        # Create gateway
-        gateway = PaymentGateway.objects.create(
-            name="Test Stripe",
-            code="stripe",
-            config=config_data
-        )
-        
-        # Refresh from database
-        gateway.refresh_from_db()
-        
-        # Verify config is accessible (decrypted automatically)
-        self.assertEqual(gateway.config, config_data)
-        self.assertEqual(gateway.config['secret_key'], "sk_test_123456789")
-    
-    def test_gateway_config_validation(self):
-        """Test gateway configuration validation"""
-        gateway = PaymentGateway(name="Test Stripe", code="stripe")
-        
-        # Test valid config
-        valid_config = {
-            "secret_key": "sk_test_123",
-            "publishable_key": "pk_test_123"
-        }
-        
-        gateway.set_config_safely(valid_config)
-        self.assertEqual(gateway.config, valid_config)
-        
-        # Test invalid config (missing required field)
-        invalid_config = {"publishable_key": "pk_test_123"}
-        
-        with self.assertRaises(ValueError):
-            gateway.set_config_safely(invalid_config)
-
-
-class AuthenticationSecurityTestCase(APITestCase):
-    """Test authentication endpoint security"""
-    
-    def setUp(self):
-        """Set up test environment"""
-        self.client = APIClient()
-        self.login_url = '/api/users/login/'
-        self.register_url = '/api/users/client-register/'
-    
-    def test_login_rate_limiting(self):
-        """Test that login attempts are rate limited"""
-        # This test would require proper rate limiting setup
-        # For now, just test that the endpoint exists and accepts requests
-        
-        response = self.client.post(self.login_url, {
-            'email': 'test@example.com',
-            'password': 'wrongpassword'
-        })
-        
-        # Should get a proper error response, not a server error
-        self.assertIn(response.status_code, [400, 401, 429])
-    
-    def test_login_input_validation(self):
-        """Test login input validation"""
-        # Test with missing fields
-        response = self.client.post(self.login_url, {})
-        self.assertEqual(response.status_code, 400)
-        
-        # Test with invalid email format
-        response = self.client.post(self.login_url, {
-            'email': 'invalid-email',
-            'password': 'password123'
-        })
-        self.assertEqual(response.status_code, 400)
-    
-    def test_registration_password_validation(self):
-        """Test registration password strength validation"""
-        # Test with weak password
-        response = self.client.post(self.register_url, {
-            'email': 'test@example.com',
-            'password': '123',
-            'confirm_password': '123',
-            'first_name': 'Test',
-            'last_name': 'User'
-        })
-        
-        self.assertEqual(response.status_code, 400)
-        # Should contain password strength feedback
-        self.assertIn('password', response.data.get('detail', '').lower())
-
-
-class WebhookSecurityTestCase(TestCase):
-    """Test webhook security improvements"""
-    
-    def setUp(self):
-        """Set up test environment"""
-        self.factory = RequestFactory()
-        self.webhook_secret = "test_webhook_secret"
-    
-    def test_webhook_signature_verification(self):
-        """Test webhook signature verification"""
-        from core.domains.communications.webhooks import verify_brevo_signature
-        
-        payload = b'{"event": "delivered", "message_id": "123"}'
-        
-        # Generate valid signature
-        expected_signature = hmac.new(
-            self.webhook_secret.encode(),
-            payload,
-            hashlib.sha256
-        ).hexdigest()
-        
-        # Test valid signature
-        self.assertTrue(verify_brevo_signature(payload, expected_signature, self.webhook_secret))
-        
-        # Test invalid signature
-        self.assertFalse(verify_brevo_signature(payload, "invalid_signature", self.webhook_secret))
-        
-        # Test missing signature
-        self.assertFalse(verify_brevo_signature(payload, None, self.webhook_secret))
-        
-        # Test missing secret
-        self.assertFalse(verify_brevo_signature(payload, expected_signature, None))
-
-
-class FileUploadSecurityTestCase(TestCase):
-    """Test file upload security"""
-    
-    def setUp(self):
-        """Set up test environment"""
-        self.user = User.objects.create_user(
-            email="admin@example.com",
-            password="AdminPass123!",
-            role="ADMIN"
-        )
-    
-    def test_file_type_validation(self):
-        """Test file type validation"""
-        from core.domains.events.services.event_services import EventFileService
-        from core.domains.events.models import Event, EventType
-        
-        # Create test event
-        event_type = EventType.objects.create(name="Test Event Type")
-        event = Event.objects.create(
-            event_type=event_type,
-            client=self.user,
-            event_date="2025-01-01",
-            guest_count=50
-        )
-        
-        # Test allowed file type (PDF)
-        pdf_content = b'%PDF-1.4 fake pdf content'
-        pdf_file = SimpleUploadedFile("test.pdf", pdf_content, content_type="application/pdf")
-        
-        file_data = {
-            'event_id': event.id,
-            'category': 'CONTRACT',
-            'name': 'Test Contract'
-        }
-        
-        # This should succeed (would need database setup to test fully)
-        try:
-            result = EventFileService.create_file(file_data, pdf_file, self.user)
-            # If we get here without exception, file validation passed basic checks
-            self.assertIsNotNone(result)
-        except Exception as e:
-            # Expected in test environment without full database setup
-            # But validation should have caught file type issues before database errors
-            self.assertNotIn("not allowed", str(e).lower())
-
-
-if __name__ == '__main__':
-    import django
-    from django.conf import settings
-    from django.test.utils import get_runner
-    
-    if not settings.configured:
-        settings.configure(
-            DEBUG=True,
-            DATABASES={
-                'default': {
-                    'ENGINE': 'django.db.backends.sqlite3',
-                    'NAME': ':memory:',
-                }
-            },
-            INSTALLED_APPS=[
-                'django.contrib.auth',
-                'django.contrib.contenttypes',
-                'core',
-            ],
-            SECRET_KEY='test-secret-key',
-        )
-    
-    django.setup()
-    TestRunner = get_runner(settings)
-    test_runner = TestRunner()
-    failures = test_runner.run_tests(["__main__"])
